refactor: 공개프로필 컴포넌트 분리 및 재사용 컴포넌트 적용
